requirements.txt
africastalking==1.2.3
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
contextlib2==0.6.0.post1
Django==3.1.7
django-environ==0.4.5
django-oidc-provider==0.7.0
djangorestframework==3.12.2
future==0.18.2
idna==2.10
psycopg2-binary==2.8.6
pycryptodomex==3.10.1
pyjwkest==1.4.2
pytz==2021.1
requests==2.25.1
schema==0.7.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
